refactor: [FE] 타입 정리
- types 폴더 내 다수의 파일 하나로 병합
- types 병합으로 인한 파일 수정
