+++ packages/client/src/hooks/useFetch.tsx
@@ -1,15 +1,16 @@
+import { AnyObject, APIMethod } from '@/types';
 import axios from 'axios';
 import { useEffect, useState } from 'react';
 
 interface Params {
   url: string;
-  method: 'get' | 'post';
-  data?: {};
+  method: APIMethod;
+  data?: AnyObject;
 }
 
 function useFetch({ url, method, data }: Params) {
   const api = process.env.REACT_APP_API_SERVER_BASE_URL;
-  const [jsonData, setJsonData] = useState<any>({});
+  const [jsonData, setJsonData] = useState<AnyObject>({});
 
   useEffect(() => {
     if (!api || !method || !url) return;
