--- packages/client/src/types/MenuList.ts
@@ -1,12 +1,0 @@
-export interface Menu {
-  id: number;
-  name: string;
-  thumbnail: string;
-  price: number;
-}
-
-export interface CafeMenu {
-  id: number;
-  cafe_name: string;
-  menus: Menu[];
-}
