--- packages/client/src/types/OrderList.ts
@@ -1,13 +1,0 @@
-export interface HistoryMenu {
-  id: number;
-  name: string;
-  price: number;
-  options: {}[];
-}
-
-export interface History {
-  id: number;
-  status: string;
-  date: string;
-  menus: HistoryMenu[];
-}
