--- packages/client/src/types/Signin.ts
@@ -1,4 +1,0 @@
-export interface chkUser {
-  code: string;
-  state: string;
-}
