--- packages/client/src/types/Signup.ts
@@ -1,5 +1,0 @@
-export interface SignupRequestBody {
-  userType: 'CLIENT' | 'MANAGER';
-  nickname: string;
-  corporate?: string;
-}
