+++ packages/client/src/utils/store.ts
@@ -1,5 +1,5 @@
 import { atom } from 'recoil';
-import { CartMenu } from '@/types/Cart';
+import { CartMenu } from '@/types';
 
 export const cartState = atom<CartMenu[]>({
   key: 'cartState',
